# src/swagger/v1/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: "Blog application"
  description: "This is a blog application where you can read many type of blogs"
servers:
  - description: "dev"
    url: "http://localhost:5000/api/v1"
...
paths:
  /auth/sign-in:
    post:
      tags:
        - auth
      description: "Api for sign in to application"
      operationId: "signIn"
      requestBody:
        description: "Provide information to login into application"
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        200:
          description: "Successfully logged in"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 400
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                  self:
                    type: string
                    example: "/auth/sign-in"
                  links:
                    type: object
                    properties:
                      articles:
                        type: string
                        example: "/articles?sortType=des&sortBy=updatedAt&limit=100&page=1&expand=comment"
                      user:
                        type: string
                        example: "/user/:id"
        400:
          $ref: "#/components/responses/400"
        500:
          $ref: "#/components/responses/500"
...
  /users:
    post:
      tags:
        - user
      description: "Create user"
      operationId: CreateUser
      parameters:
        - in: formData
          name: profile_picture
          type: file
          description: "Upload profile picture"
      requestBody:
        description: "Pass those details to create user"
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Jone deo"
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                status:
                  type: string
                  example: "pending,approved,disabled,rejected,deleted"
                role:
                  type: string
                  example: "user,admin,supper_admin"
              required:
                - name
                - email
                - password
      responses:
        201:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 201
                  message:
                    type: string
                    example: "Successfully created user account"
                  data:
                    allOf:
                      - $ref: "#/components/schemas/user"
                      - type: object
                        properties:
                          role:
                            type: string
                            example: user
                      - $ref: "#/components/schemas/timestamp"
                  self:
                    type: string
                    example: "/users"
                  links:
                    type: object
                    properties:
                      users:
                        type: string
                        example: "/users?page=1&limit=10&sortType=dsc&sortBy=updatedAt"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
    get:
      description: "Get All User"
      tags:
        - user
      operationId: "GetAllUser"
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/sortType"
        - $ref: "#/components/parameters/sortBy"
        - $ref: "#/components/parameters/search"
      responses:
        200:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: "Successfully fetched user data"
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/user"
                        - type: object
                          properties:
                            role:
                              type: string
                              example: user
                        - $ref: "#/components/schemas/timestamp"
                  self:
                    type: string
                    example: "/users?page=2&limit=10&sortType=dsc&sortBy=updatedAt"
                  links:
                    type: object
                    properties:
                      nxtPage:
                        type: string
                        example: "/users?page=3&limit=10&sortType=dsc&sortBy=updatedAt"
                      prvPage:
                        type: string
                        example: "/users?page=1&limit=10&sortType=dsc&sortBy=updatedAt"
                  pagination:
                    $ref: "#/components/schemas/pagination"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
    patch:
      tags:
        - user
      description: "Update user"
      operationId: UpdateUserPatch
      parameters:
        - in: formData
          name: profile_picture
          type: file
          description: "Upload profile picture"
      requestBody:
        description: "Pass those details to update user"
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Jone deo"
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        200:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 201
                  message:
                    type: string
                    example: "Successfully updated user account"
                  data:
                    allOf:
                      - $ref: "#/components/schemas/user"
                      - type: object
                        properties:
                          role:
                            type: string
                            example: user
                      - $ref: "#/components/schemas/timestamp"
                  self:
                    type: string
                    example: "/users"
                  links:
                    type: object
                    properties:
                      updatedUser:
                        type: string
                        example: "/user/5487"
                      userArticles:
                        type: string
                        example: "/user/5487/articles"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
    delete:
      description: "Delete User"
      tags:
        - user
      operationId: DeleteUserAccount
      responses:
        204:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 204
                  message:
                    type: string
                    example: "Successfully deleted your account"
                  self:
                    type: string
                    example: "/users"
                  links:
                    type: object
                    properties:
                      articles:
                        type: string
                        example: "/articles"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
  /user/{id}:
    get:
      description: "Get single user"
      tags:
        - user
      operationId: GetSingleUser
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        200:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 200
                  message:
                    type: string
                    example: "Successfully fetched user data"
                  data:
                    allOf:
                      - $ref: "#/components/schemas/user"
                      - type: object
                        properties:
                          role:
                            type: string
                            example: user
                      - $ref: "#/components/schemas/timestamp"
                  self:
                    type: string
                    example: "/user/585"
                  links:
                    type: object
                    properties:
                      articles:
                        type: string
                        example: "/user/585/articles?page=1&limit=10&sortType=dsc&sortBy=updatedAt"
        404:
          $ref: "#/components/responses/404"
        401:
          $ref: "#/components/responses/401"
        500:
          $ref: "#/components/responses/500"
    put:
      tags:
        - user
      description: "Update user"
      operationId: UpdateUser
      parameters:
        - in: formData
          name: profile_picture
          type: file
          description: "Upload profile picture"
        - $ref: "#/components/parameters/pathId"
      requestBody:
        description: "Pass those details to update user"
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Jone deo"
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                status:
                  type: string
                  enum:
                    - pending
                    - approved
                    - disabled
                    - rejected
                    - deleted
                role:
                  type: string
                  enum:
                    - user
                    - admin
                    - supper_admin
              required:
                - name
                - email
      responses:
        200:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 201
                  message:
                    type: string
                    example: "Successfully updated user account"
                  data:
                    allOf:
                      - $ref: "#/components/schemas/user"
                      - type: object
                        properties:
                          role:
                            type: string
                            example: user
                      - $ref: "#/components/schemas/timestamp"
                  self:
                    type: string
                    example: "/user/585"
                  links:
                    type: object
                    properties:
                      userData:
                        type: string
                        example: "/user/585"
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
    delete:
      description: "Delete User"
      tags:
        - user
      operationId: DeleteUser
      parameters:
        - $ref: "#/components/parameters/pathId"
      responses:
        204:
          description: "Success message"
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: number
                    example: 204
                  message:
                    type: string
                    example: "User delete successfully"
                  self:
                    type: string
                    example: "/user/585"
                  links:
                    type: object
                    properties:
                      users:
                        type: string
                        example: "/users"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        500:
          $ref: "#/components/responses/500"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    400:
      description: "Bad request"
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 400
              error:
                type: string
                example: "Bad request"
              data:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: "Invalid email"
    500:
      description: "Server error"
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 500
              error:
                type: string
                example: "Server error"
              message:
                type: string
                example: "Server is unable to give response. please try again later"
    401:
      description: "Unauthorized"
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 401
              error:
                type: string
                example: "Unauthorized"
              message:
                type: string
                example: "Please provide a valid authentication token."
    403:
      description: "Forbidden"
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 401
              error:
                type: string
                example: "Forbidden"
              message:
                type: string
                example: "You do not have correct access of this resource"
    404:
      description: "Not found"
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: number
                example: 204
              error:
                type: string
                example: "Not found"
              message:
                type: string
                example: "Resource not found"
  schemas:
...
    user:
      properties:
        id:
          type: number
          example: 20
        name:
          type: string
          example: "Jone Deo"
        email:
          type: string
          format: email
        profile_picture:
          type: string
          format: url
    timestamp:
      properties:
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    pagination:
      properties:
        page:
          type: number
          example: 5
        limit:
          type: number
          example: 10
        prvPage:
          type: number
          example: 4
        nxtPage:
          type: number
          example: 6
        totalPage:
          type: number
          example: 10
        totalResource:
          type: number
          example: 100
  parameters:
    page:
      in: query
      type: number
      name: page
      default: 0
      description: "Which page you want to load"
    limit:
      in: query
      type: number
      name: limit
      default: 10
      description: "How many blog you want to get at a time"
    sortType:
      in: query
      name: sortType
      schema:
        type: string
        enum:
          - asc
          - dsc
      default: dsc
      description: "How you want to sort asc or dsc"
    sortBy:
      in: query
      name: sortBy
      description: "Which field you want to use for sorting"
      type: string
      default: "updatedAt"
    search:
      in: query
      name: search
      description: "Search"
      type: string
    pathId:
      in: path
      name: id
      schema:
        type: number
      default: 1
      required: true
      description: "Pass resource id"